// src/models/product.variation.option.model.ts
export interface ProductVariationOption {
  Id?: number;
  ProductVariationId: number;
  CompanyVariationId: number;
  CompanyVariationOptionId: number;
  VariationName: string;
  OptionName: string;
  CreateDate?: string;
  CreateUserId: string;
  ModifyDate?: string;
  ModifyUserId: string;
  StatusId: number;
}